refactor: Math.ceil에 명시적으로 타입지정해서 넘기도록 수정

+++ src/main/java/wooteco/subway/domain/Fare.java
@@ -38,6 +38,6 @@ public class Fare {
     }
 
     private int calculateOverFare(int overDistance, int policy) {
-        return (int)((Math.ceil((overDistance - 1) / policy) + 1) * SURCHARGE);
+        return (int)((Math.ceil((double)((overDistance - 1) / policy) + 1)) * SURCHARGE);
     }
 }
